Added cereal headers; tweaked project settings
Using C++17 for now, will bump to 14 if we have issues on the lab
machines. Tweaked project settings so 64-bit builds also work, and added
Library/include to the include paths for shorter #includes.

If for whatever reason we decide to use boost in the future, I've also
included a powershell script that dumps boost into Library/boost and
compiles only the serialization component. This can of course be
modified to include more compiled boost components.

Note that we do not have to use cereal, but it's here if we need it.