ch - Wall/Fence aggregation and creation done
Parser takes file from https://cse125.ucsd.edu/2019/cse125g2/editor and
turns it into a vector of entities. Currently works on walls and fences
(at the moment they're the same thing, just a box).

Entities are properly generated on the server. I'm holding off on
generating the prison at the moment because we do not currently have a way
to represent it on the server. Will look into different approaches to this
in the coming days.