ch - Added collider support for non-solid objects
If state->isSolid is set to false, the object will show up in
handleCollision(), but will not be displaced by BaseCollider's
handleCollision(), and will not cause an infinite loop in the
CollisionManager. This should be useful for items, dog business, and other
such entities.

Also added an EmptyCollider class. This should be used for items that
cannot be collided with.