ch - Removed <execution> because of errors
Error is as follows:
------------------------------------------------------------------------
Error	C4996	'_Experimental_parallel_algorithms': warning STL4019:
Parallel algorithms support is experimental in this release of Visual
C++. Object files and static libraries that include <execution> may need
to be rebuilt to link with a future update of Visual C++, and programs
using <execution> may not run on Windows XP. This warning will be
removed when parallel algorithms support is completed in a future
update. You can define _SILENCE_PARALLEL_ALGORITHMS_EXPERIMENTAL_WARNING
to acknowledge that you have received this warning.
------------------------------------------------------------------------

I tried defining the variable as suggested but no dice.