ch - Game doesn't start until all clients ready
I now send a EVENT_CLIENT_READY when the client has rendered all the
entities it should have. This relies on isVisible being properly set on
all entities on the server side. If an entity is not created on the client
for a server-side entity, isVisible must be set to false. Examples include
the jail, net collider, and jail detector boxes.