ch - Game now muted during loading screen
Fixes the problem we had with the fountains being super loud at the
beginning before the player location is set. However it also mutes the
background music, which could be a problem.

Also added mechanic for dogs barking, but for some reason there are
sometimes artifcats in the sound when played for other players. Disabled
for now, will look into it if I get the chance.

Made dog yelping sound follow the dog. This results in a more intuitive
source for both the dogs and humans.